requirements.txt
aniso8601==9.0.1
anyio==3.6.1
asgiref==3.5.2
autopep8==1.6.0
backports.cached-property==1.0.2
bidict==0.22.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi==1.15.1
click==8.1.3
colorama==0.4.5
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-faker==0.2
django-formtools==2.3
django-import-export==2.8.0
django-otp==1.1.3
django-phonenumber-field==6.3.0
django-two-factor-auth==1.14.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
fake-factory==9999.9.9
fastapi==0.78.0
fastapi-socketio==0.0.9
gevent==21.12.0
graphene==3.1
graphql-core==3.2.1
graphql-relay==3.2.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
phonenumbers==8.12.51
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
python-engineio==4.3.3
python-gql==0.2.4
python-multipart==0.0.5
python-socketio==5.7.0
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
strawberry-graphql==0.114.5
tablib==3.2.1
toml==0.10.2
typing_extensions==4.2.0
tzdata==2022.1
uvicorn==0.18.2
websocket==0.2.1
websockets==10.3
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.4.0
